backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.12.1
datasets==2.16.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
aiofiles==23.2.1
# Latest FastEmbed and dependencies for advanced search
fastembed==0.4.2
sentence-transformers==3.3.1
torch>=2.0.0
transformers>=4.40.0
tokenizers>=0.19.0
huggingface-hub>=0.20.0
# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0
pytest-cov==4.1.0 